loaders in separate module, added loader registry
