+++ src/luxonis_train/core/inferer.py
@@ -19,7 +19,7 @@ from luxonis_train.utils.config import ConfigHandler
 from luxonis_train.models import Model
 from luxonis_train.models.heads import *
 from luxonis_train.utils.visualization import draw_outputs, draw_labels
-from luxonis_train.utils.loader import LuxonisLoaderTorch, collate_fn
+from luxonis_train.utils.loaders import LuxonisLoaderTorch, collate_fn
 
 
 class Inferer(pl.LightningModule):
