+++ src/luxonis_train/utils/callbacks.py
@@ -194,7 +194,7 @@ class TestOnTrainEnd(pl.Callback):
         from luxonis_ml.data import LuxonisDataset, ValAugmentations
 
         from luxonis_train.utils.config import ConfigHandler
-        from luxonis_train.utils.loader import LuxonisLoaderTorch, collate_fn
+        from luxonis_train.utils.loaders import LuxonisLoaderTorch, collate_fn
 
         cfg = ConfigHandler()
         with LuxonisDataset(
