+++ src/luxonis_train/utils/loaders/__init__.py
@@ -1,0 +1,2 @@
+from .luxonis_loader_torch import LuxonisLoaderTorch
+from .base_loader import collate_fn
